+++ TameParse/Dfa/hard_coded_symbol_table.h
@@ -51,7 +51,7 @@ namespace dfa {
         int highest = lowHigh>>8;
         
         // Return the default value if the symbol is out of range
-        if (symbol < lowest || symbol >= highest) return table[offset + 0];
+        if ((int) symbol < lowest || (int) symbol >= highest) return table[offset + 0];
         
         // Look up the symbol set
         int pos			= ((unsigned int) symbol&0xff);
